fix(all/netmovie): use Double for votes DTO field and add CLI lint rule

+++ src/all/netmovie/src/eu/kanade/tachiyomi/animeextension/all/netmovie/NetMovie.kt
@@ -481,7 +481,7 @@ data class RatingsDto(
 @Serializable
 data class RatingValDto(
     val rating: Double? = null,
-    val votes: Long? = null,
+    val votes: Double? = null,
 )
 
 @Serializable
